Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit d7a624954962def0fb780b702e09e8ffcd04ec74, reversing
changes made to f3da214de677ced022af8146c056cacb6469528e.

+++ cache/src/main/java/net/runelite/cache/definitions/MapDefinition.java
@@ -39,10 +39,10 @@ public class MapDefinition
 		public Integer height;
 		public int attrOpcode;
 		public byte settings;
-		public short overlayId;
+		public byte overlayId;
 		public byte overlayPath;
 		public byte overlayRotation;
-		public short underlayId;
+		public byte underlayId;
 	}
 
 	private int regionX;
